% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testhelper-package.R
\docType{package}
\name{assure-package}
\alias{assure}
\alias{assure-package}
\title{assure: LLM-assisted unit testing}
\description{
Supplies an addin that writes 'testthat' unit tests for selected R code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/assure}
  \item \url{https://simonpcouch.github.io/assure/}
  \item Report bugs at \url{https://github.com/simonpcouch/assure/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

}
\keyword{internal}
