% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-package.R
\docType{package}
\name{ensure-package}
\alias{ensure}
\alias{ensure-package}
\title{ensure: LLM-assisted unit testing}
\description{
Supplies an addin that writes 'testthat' unit tests for selected R code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/ensure}
  \item \url{https://simonpcouch.github.io/ensure/}
  \item Report bugs at \url{https://github.com/simonpcouch/ensure/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

}
\keyword{internal}
