% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensurer.R
\name{ensurer}
\alias{ensurer}
\title{Initialize an ensurer object}
\usage{
ensurer(
  fn = getOption(".ensure_fn", default = "chat_claude"),
  ...,
  .ns = "elmer"
)
}
\arguments{
\item{fn}{A \verb{new_*()} function, likely from the elmer package. Defaults
to \code{\link[elmer:chat_claude]{elmer::chat_claude()}}. To set a persistent alternative default,
set the \code{.ensure_fn} option; see examples below.}

\item{...}{Additional arguments to \code{fn}. The \code{system_prompt} argument will
be ignored if supplied. To set persistent defaults,
set the \code{.ensure_args} option; see examples below.}

\item{.ns}{The package that the \verb{new_*()} function is exported from.}
}
\description{
ensurers are elmer \link[elmer:Chat]{Chat()}s that know how to write testthat
unit tests. This function creates ensurers, though \code{\link[=ensure_that]{ensure_that()}} will create
ensurers it needs on-the-fly.
}
\details{
If you have an Anthropic API key (or another API key and the \verb{ensure_*()}
options) set and this package installed, you are ready to using the addin
in any R session with no setup or library loading required; the addin knows
to look for your API credentials and will call needed functions by itself.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# to create a chat with claude:
ensurer()

# or with OpenAI's 4o-mini:
ensurer(
  "chat_openai",
  model = "gpt-4o-mini"
)

# to set OpenAI's 4o-mini as the default, for example, set the
# following options (possibly in your .Rprofile, if you'd like
# them to persist across sessions):
options(
  .ensure_fn = "chat_openai",
  .ensure_args = list(model = "gpt-4o-mini")
)
\dontshow{\}) # examplesIf}
}
