% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_this.R
\name{test_this}
\alias{test_this}
\title{Write unit tests for selected code}
\usage{
test_this()
}
\value{
\code{TRUE}, invisibly.
}
\description{
This function queries an LLM to write unit tests for selected R code. To do
so, it:
\itemize{
\item Initializes a \code{\link[=test_helper]{test_helper()}}: an elmer \link[elmer:Chat]{Chat()} that knows how
to write testthat unit tests.
\item Reads the contents of the active \code{.R} file as well as the current selection.
\item Opens a corresponding test file (creating it if need be).
\item Asks the LLM to write unit tests for the current selection, using the
contents of the active \code{.R} file as context.
\item Streams the response into the corresponding test file.
}
}
